pyproject.toml
[project]
name = "bitgn-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bitgn-api-connectrpc-python>=0.9.0.1.20260324104753",
    "haystack-ai>=2.26.1",
    "ollama-haystack>=6.2.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "rich>=14.3.3",
]

[[tool.uv.index]]
url = "https://buf.build/gen/python"
